datashop-data-pipeline/requirements.txt
pandas==1.5.3
pytest==7.3.1
apache-airflow==2.6.1
python-dateutil==2.8.2
